code changes

+++ python_introduction/finance_calculator.py
@@ -2,6 +2,6 @@ monthly_income = int(input("Enter your monthly income: "))
 monthly_expenses = int(input("Enter your monthly expenses: "))
 
 monthly_savings = (monthly_income - monthly_expenses)
-psavings:int = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+psavings:float = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f'Projected savings after one year, with interest, is: ${psavings}')
